[iOS] refactor: CellInput, CommonInput 분리
